Update parameterized trigger 2.36 as opt. dep
The 2.35.1 release was in June 2017, with the 2.35.2 in September 2017.
The 2.35.1 release is used by only 1.4% of all installations.  The 2.35.2
release is 38% of the installations but is 74% of 2.35.2 installations are
on versions prior to Jenkins 2.204.1.  Since the git plugin requires at
least Jenkins 2.204.1, those 74% of installations can't use the current
git plugin anyway.

Require 2.36 as the optional dependency because over 80% of its
installations are on Jenkins 2.204.1 or newer and over 50% of all
parameterized trigger plugin installations are running 2.36 or newer.