Revise test to check system vars and Jenkins vars
Checks both system environment variables (like USER on Unix variants
and USERNAME on Windows) and Jenkins provided environment variables
(like JOB_NAME).

Asserts that the variable name does not appear in the build log.

Uses a branch based on the value of the variable and asserts that the
job is successful.  A successful job means that the branch name based
on the value of the variable was found and used in the job.

Does not assert that the specific commit from the branch is the one that
is used in the job.  That could be added.