[JENKINS-64382] Exclude JGit if submodule extension is used
The git tool chooser incorrectly selected the JGit implementation for
jobs with the submodule option enabled with no other options. It also
incorrectly selected the JGit implementation for jobs with the submodule
option enabled and recursive submodules checked.  The git client plugin
and git plugin implementations with JGit do not support submodules.  If
a repository contains submodules and the submodules are fetched, then
the git tool chooser in the git plugin should not choose JGit.

Fix the issue by always rejecting JGit as a tool if the submodule
extension is enabled.