Skip honor refspec test if env is an empty value
Fixes an array out of bounds when the environment variable being tested
is an empty string.  The test is checking that the variable is correctly
expanded in the refspec but is not designed to test the case where the
string is empty.  It is possible to create a refspec that would work
with the empty string value of the environment variable, but it is more
complicated and does not add value to the scenarios evaluated by the test.

The plugin bom tests of the most recent git plugin release are failing
because the plugin bom test environment has the value of the environment
variable `USER` as the empty string.