[JENKINS-64720] Fix docs to not reference tag checkout by git step
The git Pipeline step is a simplified shorthand for a subset of the
more powerful checkout step.

The checkout step is the preferred SCM checkout method. It provides
significantly more functionality than the git step.

The documentation previously stated that tag checkout is not supported
in one list but then suggested a tag name was a valid branch name
elsewhere in the same help text.  In the arguments help text it
sugghested a tag name was a valid branch name and did not state that
tag names are not supported as a branch name in the `git` Pipeline
step.