Add deprecation message for submodule config
The doGenerateSubmoduleConfigurations and submoduleCfg settings are
not available from Freestyle projects, matrix projects, or other
legacy projects.  They are available from the Pipeline Syntax Editor,
the configuration as code plugin, and the JobDSL plugin.  They are
intentionally undocumented and will remain undocumented except to declare
that they will be removed in a future release.

The settings are not tested in anything other than their default values
(doGenerateSubmoduleConfigurations: false and submoduleCfg: []).

Set expectations that user provided values will be ignored in a future
release.