Add Pipeline Syntax video segment to docs
Use pipeline syntax image that is a hyperlink to a 90 second video on
Pipeline Syntax.  Many feedback comments on the documentation pages are
seeking examples when I believe that they could use the Pipeline Syntax
helper and obtain most of the results they need.

Could not find a way to embed a video player in the ASCIIDoc file that
would display on GitHub.  Embedding the image and making it a hyperlink
to the YouTube video is a first experiment to see if the embedded video
is considered helpful by users.