[JENKINS-64712] Fix unexpected scrolling of nested checkbox
Since the GitTool data model has the useExistingAccountWithSameEmail
and the useExistingAccountWithSameEmail fields at the same level, this
simplifies the user experience by displaying them both at the time.

Existing values are loaded as expected on startup.

Values are persisted to the GitTool.xml configuration file when the
configuration is applied or saved.

Appears to be purely a presentation convenience to only show the
useExistingAccountWithSameEmail if the useExistingAccountWithSameEmail
checkbox was enabled.

Tested interactively with Jenkins 2.263.4, Jenkins 2.277.1-rc, and
Jenkins 2.281.  Behaves as expected in all cases.

Hints in other communities suggest that the unexpected scrolling is
caused by specific CSS markup, but I can't find that specific markup
in my search of the generated page.  Rather than fight with CSS markup,
it is easier to move the nested checkbox to the top level.