Resolving gitTool return error due to assumed index for ToolInstallation array
The method method signature and implementation have been changed to return Cli Git specific git tool. The methods name have been changed to `getCliGitTool`.

Instead of returning a generic git tool as done by resolveGitTool as it won't sever our entire purpose the `getCliGitTool` method is returning a git tool which is both cli git type specific and name specific provided by the user.

This method ensures that the gitTool available to the class implementing this interface is of type cli git tool.