do not run commit hooks
if you are mandated to have something like `git secrets` installed then
the test fails as it takes more time that the test allows to make the
commit.

On windows git secrets can easily make a commit take 20s to complete or
more (but a lower bound is around 5 seconds).

As we are doing timing here and we do not care about commit hooks, we
can just skip them so they do not interfere.