Add GIT_BRANCH token macro tests for Pipeline
Test current behavior, even when the current behavior is unexpected.

The token macro documentation for the git plugin's GIT_BRANCH variable
indicates that fullName=false should leave only the basename of the
branch, but the tests show that in many cases, the full name of the
branch remains.

Since Pipeline has all the string manipulation of groovy, there is a
workaround that does not involve using the token macro facility.