Do not run tests in parallel
James Nord recommends that parallel testing be the choice of the user
and that it be run through an argument to the mvn command rather than
configuring it in the pom file.

Since there is a report that the tests fail when forked on Windows,
let's follow the recommendation from James and require the user to
choose to run parallel tests by command line argument.

I don't see the test failure consistently on my Windows computers, but
have seen it at least once on them.