Avoid command processor expansion of password values
Windows command processor treats '&' as a special character in
command lines.  That makes the sample password '&Ampersand&' expand
incorrectly in the test code that was using `echo` on Windows.

Instead of using `echo` to check that the password is being masked,
use the Windows `type` command to display the contents of the auth.txt
file contents.

This simplifies the test case while still checking that the masking
is performed as expected.