[JENKINS-67981] handle any AuthenticationException
When processing the changelog, if a looked user is disabled in LDAP the
git plugin would throw an `UncheckedExecutionException` with the cause
being `DisabledException`. The unhandled exception would cause a build
failure whenever processing a commit that is determinated to be from a
disabled user.

Jenkins `UserDetailsCache.loadUserByUsername()` receives an
`ExecutionException` which wraps the `AuthenticationException`. The
method inspects the cause, handles instances of
`UsernameNotFoundException` which it throws but ignores any other
`AuthenticationException`.

This introduces a `User` getter to handle any `AuthenticationException`
and throws it instead of the `ExecutionException`. Callers can thus:

  try {
    user = getUser("someOne");
  } catch (AuthenticationException e) {
    return User.getUnknown();
  }
  // Do something with "user"

I have added a test to ensure a `DisabledException` does yield an
unknown user. It only test a single code path though.

Reference: https://phabricator.wikimedia.org/T315897