Fix NPE in Windows test of GitHooksTest
Also intentionally avoids all calls to Functions.isWindows because that
method is not used in any other location in the tests or in the
production code.  It is not used because there was a time when a remote
class loader issue was detected due to loading the Hudson.Functions class
just to decide if a class was running on Windows.  The inline test is
simple and is easy to read.

Confirmed once again that the GitHooksTest works in my dev env.