Remove wasted space on Repository URL field
The repeatable delete button in the groovy form should be consistent
with the repeatable delete button used in jelly files.  No need to add
the extra div and waste vertical space.

The "show-if-not-only" class does not seem to have any affect on the
shipping version of the plugin with Jenkins 2.361.2.  I can delete the
Repository URL region because the delete "x" is displayed even when the
entry is the only entry of that type.  This change does not make the
current condition any worse with regards to the existence of the delete
"x" in the top right corner of the field and makes it more consisistent
with the improvement from @janfaracik.