Do not assert changeset paths are unique
Previously, the code asserted that the paths in a changeset were unique.
They may be unique when a single commit is in the changeset.  In those
cases where the changeset includes a merge, the paths are not unique if
both sides of the merge modified a file.  When a file is modified on
both sides of a merge, then the file will be correctly listed on both
sides of the merge.