Remove credentials "Add" button workaround (#1374)
* [JENKINS-70215] Remove bad credentials button markup

https://issues.jenkins.io/browse/JENKINS-70215 reports that the "Add
credentials" button has unexpected lines drawn inside the button on
Google Chrome.  The position of the lines inside the button seem to
depend on the location of the position of the button on the screen.

https://issues.jenkins.io/browse/JENKINS-19124 was listed in the
comments as the reason for the embedded JavaScript code in that button
definition.  The Mercurial plugin does not use that JavaScript code in
its credentials button.  It seems safe to remove it from the git
plugin "Add" button for credentials.

Interactive testing before the change confirmed that the unexpected
line(s) were drawn inside the button.  Interactive testing after the
change confirmed that the unexpected line(s) were no longer being
drawn inside the button.

* Replace groovy form with jelly

Much more attention is being given to jelly forms than to groovy forms.
Replace the groovy form with a jelly form because it is better for the
future of the plugin.  This is a behavior preserving transformation,
expected to cause no change in behavior.