Test workflow snippet generation for checkout step
Workflow snippets should not display parameters if the default value
of the parameter is unmodified.

The `git` step does that very well.  When a credentialId is not provided,
the snippet does not include a parameter for the credentialId.

The `checkout` step does not do that very well.  Many default parameters
are included in the generated snippet.

This test shows the issue without providing any resolution for the
issue.  The issue is shown by constructing an SCM Pipeline step and
then asserting it has the expected contents with the extra "junk".
An additional assertion is made that the Pipeline definition parses to
the same result even without the extra "junk".