Limit test duration of the longest running tests
Use JUnit assume to skip tests after time exceeded

JUnit assume will flag the test skip as a warning in the Jenkins job.
That's a good thing in this case, since we want to know when tests are
skipped due to excessive elapsed test time.

Run tests in random order

Run all tests when not in CI

Report duration limit in skip message

Limit GitSCMTest to not more than 9.5 minutes

Limit GitStatusTest to not more than 5 minutes

Limit GitSCMSlowTests to not more than 3 minutes

Limit AbstractGitSCMSourceTest to not more than 2.5 minutes

Limit GitToolChooserTest to not more than 2.5 minutes