Split GitStatusTest to place theory tests in new class
Resolves a test failure when the theory based test exceeds expected
maximum run time.  Maven surefire support for JUnit theories appears
to be limited.  JUnit ongoing support of theories seems to be low to
non-existent.  Removes the more complicated theory based tests into a
separate class so that they can eventually be replaced with parameterized
tests or another simpler, more widely supported form.

A comment on stackoverflow says that theories are not supported in
JUnit 5.  That provides another reason for the eventual removal of theory
based tests.

https://stackoverflow.com/questions/51543612/junit-5-theory-and-datapoint