Add symbols for easier Pipeline authoring
Pipeline authoring with the git plugin has not been using the symbol
facility that is used elsewhere in Jenkins.

Previously, a Pipeline checkout might look like this:

    checkout([$class: 'GitSCM',
              userRemoteConfigs: [[url: 'https://github.com/jenkinsci/git-plugin.gi' ]],
              branches: [[name: 'master']],
              extensions: [[$class: 'CloneOption', noTags: true, timeout: 3],
                           [$class: 'LocalBranch', localBranch: branch],
                          ],
             ])

With the addition of symbols, the earlier Pipeline checkout continues
to be supported, but it can also be written as:

    checkout scmGit(
              userRemoteConfigs: [[url: 'https://github.com/jenkinsci/git-plugin.gi' ]],
              branches: [[name: 'master']],
              extensions: [cloneOption(noTags: true, timeout: 3),
			   localBranch('master')
			  ],
             )

The symbols make it easier to read and easier to understand.

Change includes updated documentation with Pipeline examples and
screenshots taken with Jenkins 2.375.1.

Thanks to Karl Shultz for starting the work on this improvement.