Require Jenkins 2.387.3 or newer
Require Jenkins 2.387.3 or newer so that users do not think that versions
older than 2.387.3 are actively tested.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

Choosing a Jenkins version to build against recommends 2.387.3 or 2.375.4
as the base version.  Versions prior to 2.375.4 have known security
vulnerabilities in the plugin manager.  I no longer test anything older
than 2.387.3.

As of May 1, 2023 78% of the 96k installations of the most recent release
(5.0.0, Dec 30, 2022) are running on Jenkins 2.387.1 or newer.