Fix spotbugs warning on repo cache size updates
Spotbugs correctly reported a race condition where the check for
insertion of a value might report the value is missing then another
thread inserted the value before the current thread performed the
`put`.  Simpler to `put` the value every time the method is called and
then check for the rare case when the value that was put is smaller
than the value that was there perviously.

Repository size cache is used as a hint.  Errors in the size cache may
lead to suboptimal choices temporarily, but they should not lead to
incorrect behavior.