Fix erratic GitStatusTest (#1480)
* Fix random GitStatusTest failures

Test already had code that waits for jobs to complete on Windows.
The improved job leak detection that was implemented in the Jenkins
test harness will sometimes show the same conditions on Unix builds.
Wait for jobs to complete on Windows and on Unix.

* Remove non JenkinsRule tests to simplify cleanup

* Move simple GitStatusTest items to new class

No need to complicate GitStatusTest with the test methods that do not
require a JenkinsRule.  Use a simpler test structure and a simpler test
file for the GitStatusTest assertions that do not require a JenkinsRule.

* Assert that the "all" view must exist

* Report a message if no jobs found for wait loop

Not an error, but surprising based on my testing.  None of my test
machines are fast enough to complete all the jobs without needing any
time for waiting.

* Replace Hudson with Jenkins in comments

* Remove unused imports