Require Jenkins 2.401.3 or newer
Git client plugin will soon require 2.401.3 or newer so that it can
depend on the gson-api plugin rather than bundling the gson-api jar
file as a transitive dependency for the Eclipse JGit LFS library.

42% of all installations of the git plugin are already running Jenkins
2.401.3 or newer.

75% of all installations of git plugin 5.0.0 or newer are already
running Jenkins 2.401.3 or newer.  The git plugin 5.0.0 release was
more than 11 months ago.

94% of the 107000 installations of git plugin 5.2.0 + 5.2.1 are
running on Jenkins 2.401.3 or newer.

The majority of git plugin users that are upgrading their plugin
version are upgrading to releases of the plugin on Jenkins 2.401.3 or
newer.

Security advisories that apply to Jenkins 2.401.2 or 2.387.3 and do
not apply to Jenkins 2.401.3 include:

* https://www.jenkins.io/security/advisory/2023-07-26/
* https://www.jenkins.io/security/advisory/2023-06-14/