Replace multiple images with a single image
The multi-image layout renders correctly on github.com but does not
render correctly on plugins.jenkins.io.  The technique that was being
used is uncommon for Jenkins plugins.  Much better to use the layout of
a single image that is approximately the same width as the other images
included in the page.