Skip testMergeWithMatrixBuild if gpgsign enabled
The testMergeWithMatrixBuild test fails randomly on several machines when
commit.gpgsign and tag.gpgsign are not enabled if the TestPreBuildMerge
implementation is used. It passes consistently when PreBuildMerge is
used.

Rather than spend the time trying to diagnose the intermittent failures,
this configuration allows the test to be skipped if either of those
configuration settings are enabled.

Other tests in this class are able to use TestPreBuildMerge without issue.

Tested by enabling commit.gpgsign in ~/.gitconfig and by enabling
commit.gpgsign in a $XDG_CONFIG_HOME/.gitconfig file.  Test was skipped
when gpgsign was enabled in either of those files.