Add issue comment to promoted builds dependabot exclusion
https://github.com/jenkinsci/bom/pull/3172 shows that the promoted build
plugin dependency in the git plugin needs to be updated in the same
plugin bill of materials release as the parameterized trigger plugin.
Link to the GitHub issue in the plugin BOM that will move the management
of the promoted builds plugin version to the plugin BOM, instead of
managing the version in the specific plugins.  That move will also help
the Apache Maven integration plugin.