Downgrade matrix-project optional dependency for PCT
The plugin compatibility tester does not load the version of the optional
dependency that is specified by the plugin (git plugin in this case)
when loading plugins to test a plugin that depends on the git plugin
(pipeline-model-definition in this case).  Downgrading this optional
dependency should be harmless to other plugins and to users and may
allow the plugin compatibility tester to pass tests in the plugin bill
of materials.