[JENKINS-43052] Warn in Pipeline build log if a deprecated extension is used
When extensions created for freestyle projects have a more general
purpose replacement in Pipeline, the more general purpose Pipeline step
is preferred.  This method allows the deprecated extension to suggest
the preferred alternative.

When one of those extensions is used in a Pipeline, write a
***DEPRECATED** warning message to the build log so that the user can
eventually remove the use of the deprecated extension from their Pipeline.

The deprecated warning message in the build log does not alter any
behavior, it only provides a message for the user.

Suggest alternatives for those cases where a reasonable alternative exists.

Tested by running a Pipeline job with one of the deprecated extensions
and confirmed that the message was displayed as expected.

Tested by running a freestyle job with the same extension and confirmed
that no message was written to the build log.