Add sparseCheckout symbol for Pipeline (#1638)
Includes a test that confirms the sparseCheckout syntax works as expected.

Tested interactively to confirm that the new symbol behaves the same as
the old syntax behaved.  Does not include detailed functional testing
that the sparseCheckout symbol works as expected, since there are already
tests of sparse checkout in other tests.