Fix acceptance tests bny reverting capitalization change
The Jenkins acceptance test harness uses the display name to locate UI
components and test them.  It locates those components by display name.

The tests in the acceptance test harness are especially valuable now while
we are preparing for the Spring Security 6.x upgrade.  It is better to
retain the previous capitalization and preserve Jenkins acceptance test
harness results rather than need to spread the capitalization change
into the acceptance test harness and related components.

Added TODO item to fix the capitalization in both the git plugin and
the acceptance test harness in a future release.