Temporarily depend on latest mina sshd plugin
https://github.com/jenkinsci/mina-sshd-api-plugin/releases/tag/2.14.0-133.vcc091215a_358
includes a bug fix to move fips-bundle-test to test scope and to not
include commons-io in the dependencies of the plugin.  That release is
not yet available in plugin BOM.  Once it is available in plugin BOM,
this workaround can be removed.

https://github.com/jenkinsci/mina-sshd-api-plugin/pull/119 is the pull
request that resolved the issue in mina sshd api plugin.