[JENKINS-75162] Fix fragment identifiers to plugins.jenkins.io content
The previous fragment identifiers in the documentation links no longer
exist in the plugins.jenkins.io page content.  Use fragment identifiers
that exist so that a click of the link takes the reader to the correct
location in the page content.

Partial fix of https://issues.jenkins.io/browse/JENKINS-75162