Upgrade matrix-project test dependency based on plugin BOM results
The 2.452.x plugin bill of materials has reached end of life so this
workaround can be upgraded to a newer version of matrix-project plugin.

Retain the workaround because the pipeline model definition plugin fails
to start on the 2.479.x and 2.492.x lines if the workaround is removed.

Testing done:

Confirmed that the plugin BOM tests pass with a pre-release

  export PLUGINS=pipeline-model-definition
  export TEST=InjectedTest
  LINE=2.462.x bash local-test.sh
  LINE=2.479.x bash local-test.sh
  LINE=2.492.x bash local-test.sh
  LINE=weekly  bash local-test.sh