Use workspace polling for empty branches
When a single empty branch is encountered, getSingleBranch converts it
to the ** wildcard spec like BranchSpec. Since that's not null,
requiresWorkspaceForPolling says it can be used with remote polling.
However, remote polling doesn't work well for wildcard branches since it
doesn't have access to the remote history to prune old branches. Detect
this special case and require workspace polling for it.

I'm not sure why getSingleBranch goes to such effort to reject wildcard
branches but then treats allows an empty branch as a wildcard. It seems
like it should return null for an empty branch, but it also gets passed
to getCandidateRevisions, which may expect those semantics.