Remove incorrect annotation on GitSCM.getBrowser()
Jesse Glick notes in https://issues.jenkins.io/browse/JENKINS-75288:

> adding @Whitelisted to a method already marked @Override is clearly
> a mistake—this would never have any effect.

I didn't realize that was a mistake when I added the annotation.
Let's remove the annotation so that the code does not mislead someone
to think that this can be used from a Pipeline without changes to
hudson.scm.SCM.getBrowser().

In a multibranch Pipeline, the default checkout assigns current values
to the scm object, including scm.browser.  That value can't be read
unless the hudson.scm.SCM.getBrowser() method is approved through script
security.