Add tracking pixel to documentation page (#1777)
Collect viewing data for the Git plugin documentation so that it can
be better tuned to the needs of its readers.  Identify if there are
specific countries that are viewing the page so that we might consider
a translation into the primary language of that country.

Use analytics service from scarf.sh as provided through the Linux
Foundation.