[JENKINS-75854] Do not fetch tags in multibranch Pipeline if "Discover tags" is not enabled (#1776)
* Do not fetch tags if not needed

If the "Discover Tags" trait is not set, there is no need to fetch tags
on branch indexing.

* Add a test

* Correct error in test - refs/tags/* not retrieved without tag discovery

The refs/tags namespace should not have been retrieved previously when
tag discovery was not enabled.

* Empty commit to trigger a build
* Clear org.jenkinsci.plugins.gitclient.Git.mockClient after the test
* Move wantsTags test to new class
* Move wantTags test into a separate file

Test the heads that are returned in combinations of traits.

Use Before and After methods to register the MockGitClient so that we
can avoid a try / finally block in the test.

* Add an escape hatch to retain bug compatibility (just in case)

Users may find that the bug fix has a negative impact on their environment
and may not be able to immediately adjust their multibranch Pipeline
definitions.  Accept a system property that restores the buggy behavior.

* Split wantTags tests into separate tests for coverage and clarity

Tests the impact of GitSCMSource.IGNORE_TAG_DISCOVERY_TRAIT on the
behavior.

* Add javadoc for IGNORE_TAG_DISCOVERY_TRAIT

Explain https://issues.jenkins.io/browse/JENKINS-75854

https://github.com/jenkinsci/git-plugin/pull/1776

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>