Fix formatting of scmGit Pipeline help
Was displaying red text for most of the page at

https://www.jenkins.io/doc/pipeline/steps/params/scmgit/

The <code> tag that opened on the checkout keyword was not closed.
Close that tag so that the page renders correctly in the Pipeline steps
reference on the Jenkins documentation site.

The Pipeline steps reference inside the running Jenkins controller was
not affected.  It seems to ignore the <code> tags.

Testing done:

Code review confirms that the closing and opening tags are matched.