Test with Java 25 and Java 21 (#1826)
Java 25 released in September 2025 and the Jenkins project wants
to support it soon after its release.  This pull request updates the
plugin to compile and test with Java 25 and Java 21.  The Java byte code
continues to be Java 17 byte code from both those compilers due to the
configuration of the plugin pom.  That is intentional because we continue
to support Java 17 but we don't need to compile and test with Java 17.
We've seen no issues that were specific to Java 17.

Testing done:

* Confirmed that tests are successfil with Java 25